+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - switch all files from `js` to `ts`/`tsx` and set better types
 - add `onDragStart` and `onDragEnd` props
 - make plus/minus icons to be inline in the default css file
+- add `disableCollapse` and `disableDrag` props
 
 ## 2.0.0